pyproject.toml
[tool.poetry]
name = "ebookspeech"
version = "0.1.0"
description = "Yet another openai TTS"
authors = ["Agata Ordano <gh@choccymilk.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.40.6"
python-dotenv = "^1.0.1"
typer = {extras = ["all"], version = "^0.12.3"}
pypandoc = "^1.13"
pydub = "^0.25.1"
tiktoken = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
